Update dg2fa.sh
更新13-31

+++ dg2fa.sh
@@ -13,19 +13,24 @@ case $option in
         # 修改主机名
         read -p "修改主机名，方便区分记忆: " new_hostname
         hostnamectl set-hostname "$new_hostname"
+        sed -i "1i 127.0.0.1       $new_hostname" /etc/hosts
         echo "主机名修改完成。"
 
         # 更新软件包列表并安装Google Authenticator和ntpdate
         echo "更新软件包列表并安装Google Authenticator和ntpdate..."
         apt update && apt install -y libpam-google-authenticator ntpdate
-        echo -e "Google Authenticator安装完成，下面即将配置验证器。\n配置验证器时会有多次询问，请详细查看：\n1.是否启用基于时间的一次性密码验证，建议选择 y，然后使用谷歌验证器扫描二维码。\n2.是否在用户目录下更新验证文件，建议选择 y\n3.是否禁止一个口令多次使用，防止中间人攻击，建议选择 y\n4.【最长的一段】是否延长口令验证时间，避免客户端与服务器时间误差，除非时间同步很差，否则建议选择 n\n5.是否限制尝试次数，防止暴力攻击，建议选择 y"
-
         # 修改时区为上海并同步时间
+        read -p "是否要修改时区并同步时间？(y/n，默认为y): " response
+        if [ "$response" = "y" ] || [ -z "$response" ]; then
         echo "修改时区为上海并同步时间..."
         ln -sf /usr/share/zoneinfo/Asia/Shanghai /etc/localtime
         ntpdate time.nist.gov
         echo "时区和时间同步完成。"
-
+        else
+        echo "跳过修改时区和同步时间。"
+        fi
+        # 安装前说明
+        echo -e "Google Authenticator安装完成，下面即将配置验证器。\n配置验证器时会有多次询问，请详细查看：\n1.是否启用基于时间的一次性密码验证，建议选择 y，然后使用谷歌验证器扫描二维码。\n2.是否在用户目录下更新验证文件，建议选择 y\n3.是否禁止一个口令多次使用，防止中间人攻击，建议选择 y\n4.【最长的一段】是否延长口令验证时间，避免客户端与服务器时间误差，除非时间同步很差，否则建议选择 n\n5.是否限制尝试次数，防止暴力攻击，建议选择 y"
         # 配置Google Authenticator
         echo "配置Google Authenticator..."
         google-authenticator
